requirements.txt
Django==1.8
django-model-utils==2.6.1
